+++ src/main/java/fr/saddem/bank/entity/DepositOperation.java
@@ -1,9 +1,10 @@
 package fr.saddem.bank.entity;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class DepositOperation extends Operation{
-    public DepositOperation(String id, LocalDateTime operationDate, Double amount, Double balance, Account account) {
+    public DepositOperation(String id, LocalDateTime operationDate, BigDecimal amount, BigDecimal balance, Account account) {
         super(id, operationDate, amount, balance, account);
     }
 }
