+++ src/main/java/fr/saddem/bank/service/AccountService.java
@@ -1,5 +1,6 @@
 package fr.saddem.bank.service;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Optional;
 
@@ -10,7 +11,7 @@ import fr.saddem.bank.exceptions.NotEnoughBalanceException;
 
 public interface AccountService {
     public Optional<Account> findAccountById(Long idAccount) throws AccountNotFoundException;
-    public void deposit(Long idAccount, Double amount) throws AccountNotFoundException;
-    public void withdrawl(Long idAccount, Double amount) throws AccountNotFoundException, NotEnoughBalanceException;
+    public void deposit(Long idAccount, BigDecimal amount) throws AccountNotFoundException;
+    public void withdrawl(Long idAccount, BigDecimal amount) throws AccountNotFoundException, NotEnoughBalanceException;
     public Optional<ArrayList<Operation>> getOperationsByAccountId(Long idAccount) throws AccountNotFoundException;
 }
